+++ Source/OxyPlot.ImageSharp.Tests/PngExporterTests.cs
@@ -4,15 +4,15 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace OxyPlot.SkiaSharp.Tests
+namespace OxyPlot.ImageSharp.Tests
 {
     using System;
     using System.IO;
-    using ExampleLibrary;
     using NUnit.Framework;
-    using OxyPlot.Annotations;
+
     using OxyPlot.Series;
-    using OxyPlot.SkiaSharp;
+    using OxyPlot.ImageSharp;
+    using OxyPlot.Annotations;
 
     [TestFixture]
     public class PngExporterTests
@@ -30,7 +30,7 @@ namespace OxyPlot.SkiaSharp.Tests
         [Test]
         public void Export_SomeExamplesInExampleLibrary_CheckThatAllFilesExist()
         {
-            var exporter = new PngExporter { Width = 400, Height = 300 };
+            var exporter = new PngExporter(400, 300);
             var directory = Path.Combine(this.outputDirectory, "ExampleLibrary");
             ExportTest.Export_FirstExampleOfEachExampleGroup_CheckThatAllFilesExist(exporter, directory, ".png");
         }
@@ -39,7 +39,7 @@ namespace OxyPlot.SkiaSharp.Tests
         public void ExportToStream()
         {
             var plotModel = CreateTestModel1();
-            var exporter = new PngExporter { Width = 400, Height = 300 };
+            var exporter = new PngExporter(400, 300);
             var stream = new MemoryStream();
             exporter.Export(plotModel, stream);
 
@@ -62,7 +62,7 @@ namespace OxyPlot.SkiaSharp.Tests
             var plotModel = CreateTestModel1();
             plotModel.Background = OxyColors.Yellow;
             var fileName = Path.Combine(this.outputDirectory, "Background_Yellow.png");
-            var exporter = new PngExporter { Width = 400, Height = 300 };
+            var exporter = new PngExporter(400, 300);
             using (var stream = File.OpenWrite(fileName))
             {
                 exporter.Export(plotModel, stream);
@@ -73,6 +73,7 @@ namespace OxyPlot.SkiaSharp.Tests
 
         [Test]
         [TestCase(0.75)]
+        [TestCase(1)]
         [TestCase(1.2)]
         [TestCase(2)]
         [TestCase(3.1415)]
@@ -84,46 +85,11 @@ namespace OxyPlot.SkiaSharp.Tests
             Directory.CreateDirectory(directory);
 
             var fileName = Path.Combine(directory, $"Resolution{resolution}.png");
-            var exporter = new PngExporter { Width = (int)(400 * factor), Height = (int)(300 * factor), Dpi = resolution };
+            var exporter = new PngExporter((int)(400 * factor), (int)(300 * factor), resolution);
 
             using (var stream = File.OpenWrite(fileName))
             {
                 exporter.Export(plotModel, stream);
-            }
-
-            Assert.IsTrue(File.Exists(fileName));
-        }
-
-        [Test]
-        [TestCase(true)]
-        [TestCase(false)]
-        public void ExportUseTextShapingAlignment(bool useTextShaping)
-        {
-            var model = RenderingCapabilities.DrawTextAlignment();
-            model.Background = OxyColors.White;
-            var fileName = Path.Combine(this.outputDirectory, $"Alignment, UseTextShaping={useTextShaping}.png");
-            var exporter = new PngExporter { Width = 450, Height = 200, UseTextShaping = useTextShaping };
-            using (var stream = File.OpenWrite(fileName))
-            {
-                exporter.Export(model, stream);
-            }
-
-            Assert.IsTrue(File.Exists(fileName));
-        }
-
-        [Test]
-        [TestCase(true)]
-        [TestCase(false)]
-        public void ExportUseTextShapingMeasurements(bool useTextShaping)
-        {
-            var model = RenderingCapabilities.DrawTextWithMetrics("TeffVAll", "Arial", 60, double.NaN, double.NaN, 105, double.NaN, double.NaN, double.NaN, double.NaN, double.NaN, "");
-
-            model.Background = OxyColors.White;
-            var fileName = Path.Combine(this.outputDirectory, $"Measurements, UseTextShaping={useTextShaping}.png");
-            var exporter = new PngExporter { Width = 450, Height = 150, UseTextShaping = useTextShaping };
-            using (var stream = File.OpenWrite(fileName))
-            {
-                exporter.Export(model, stream);
             }
 
             Assert.IsTrue(File.Exists(fileName));
@@ -162,7 +128,7 @@ namespace OxyPlot.SkiaSharp.Tests
             plotModel.Annotations.Add(imageAnnotation);
 
             var fileName = Path.Combine(this.outputDirectory, $"PlotBackground{(interpolate ? "Interpolated" : "Pixelated")}.png");
-            var exporter = new PngExporter { Width = 400, Height = 300 };
+            var exporter = new PngExporter(400, 300);
             using (var stream = File.OpenWrite(fileName))
             {
                 exporter.Export(plotModel, stream);
@@ -204,7 +170,7 @@ namespace OxyPlot.SkiaSharp.Tests
             plotModel.Annotations.Add(imageAnnotation);
 
             var fileName = Path.Combine(this.outputDirectory, $"LargeImage{(interpolate ? "Interpolated" : "Pixelated")}.png");
-            var exporter = new PngExporter { Width = 400, Height = 300 };
+            var exporter = new PngExporter(400, 300);
             using (var stream = File.OpenWrite(fileName))
             {
                 exporter.Export(plotModel, stream);